% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{annoInfo}
\alias{annoInfo}
\title{calculate percentages, fold enrichment and significance of an annotation}
\usage{
annoInfo(anno, df, ref)
}
\arguments{
\item{anno}{string}

\item{df}{data.frame}

\item{ref}{data.frame}
}
\value{
data.frame
}
\description{
calculate percentages, fold enrichment and significance of an annotation
}
