% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{celLoc}
\alias{celLoc}
\title{Extract cellular localization information}
\usage{
celLoc(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Extract cellular localization information
}
