% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{cysFun}
\alias{cysFun}
\title{Extract cysteine functional annotations}
\usage{
cysFun(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Extract cysteine functional annotations
}
