% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\name{enrichAll}
\alias{enrichAll}
\title{Enrichment analysis including GO, KEGG, Reactome}
\usage{
enrichAll(poi, u, Org)
}
\arguments{
\item{poi}{vector}

\item{u}{vector}

\item{Org}{string}
}
\value{
list
}
\description{
Enrichment analysis including GO, KEGG, Reactome
}
