% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human}
\alias{human}
\title{human data based on UniProt}
\format{
A data frame with 20371 rows
}
\source{
\url{http://uniprot.org/proteomes/UP000005640}
}
\usage{
human
}
\description{
human data based on UniProt
}
\keyword{datasets}
