% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanAnno}
\alias{humanAnno}
\title{Cys functional annotation of human gene}
\format{
A data frame with 20371 rows and 15 variables:
}
\source{
\url{http://uniprot.org/proteomes/UP000005640}
}
\usage{
humanAnno
}
\description{
Cys functional annotation of human gene
}
\keyword{datasets}
