% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanPPI}
\alias{humanPPI}
\title{degrees of human proteome}
\format{
A data frame with 20371 rows and 3 variables:
}
\source{
\url{http://uniprot.org/proteomes/UP000005640}
}
\usage{
humanPPI
}
\description{
degrees of human proteome
}
\keyword{datasets}
