% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanUnfold}
\alias{humanUnfold}
\title{IUPred2A prediction of human proteome}
\format{
A data frame with 20371 rows and 3 variables:
}
\source{
\url{http://uniprot.org/proteomes/UP000005640}
}
\usage{
humanUnfold
}
\description{
IUPred2A prediction of human proteome
}
\keyword{datasets}
