% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUPred2A.R
\name{mergeFasta}
\alias{mergeFasta}
\title{Merge and summarize IUPred results}
\usage{
mergeFasta(inputDir = utils::choose.dir())
}
\arguments{
\item{inputDir}{path}
}
\description{
Merge and summarize IUPred results
}
