% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse}
\alias{mouse}
\title{mouse data based on UniProt}
\format{
A data frame with 17068 rows
}
\source{
\url{http://uniprot.org/proteomes/UP000000589}
}
\usage{
mouse
}
\description{
mouse data based on UniProt
}
\keyword{datasets}
