% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseAnno}
\alias{mouseAnno}
\title{Localization and Cys functional annotation of mouse gene}
\format{
A data frame with 17068 rows and 15 variables:
}
\source{
\url{http://uniprot.org/proteomes/UP000000589}
}
\usage{
mouseAnno
}
\description{
Localization and Cys functional annotation of mouse gene
}
\keyword{datasets}
