% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mousePPI}
\alias{mousePPI}
\title{degrees of mouse proteome}
\format{
A data frame with 17068 rows and 3 variables:
}
\source{
\url{http://uniprot.org/proteomes/UP000000589}
}
\usage{
mousePPI
}
\description{
degrees of mouse proteome
}
\keyword{datasets}
