% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plots}
\usage{
plotVolcano(df, fcTH = 1.5, fdr = 0.05)
}
\arguments{
\item{df}{data.frame}

\item{fcTH}{double}

\item{fdr}{double}
}
\value{
ggplot2.object
}
\description{
Volcano plots
}
