% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppiProteome.R
\name{ppiProteome}
\alias{ppiProteome}
\title{export the degree and betweenness of the proteome}
\usage{
ppiProteome(species, score_threshold = 600)
}
\arguments{
\item{species}{string}

\item{score_threshold}{int}
}
\value{
data.frame
}
\description{
export the degree and betweenness of the proteome
}
