% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess proteomics data}
\usage{
preprocess(df, left, right)
}
\arguments{
\item{df}{dataframe}

\item{left}{string}

\item{right}{string}
}
\value{
dataframe
}
\description{
Preprocess proteomics data
}
