% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process2groups}
\alias{process2groups}
\title{2 experimental conditions}
\usage{
process2groups(df, h0, h1, seperator = "-1")
}
\arguments{
\item{df}{data.frame}

\item{h0}{string}

\item{h1}{string}

\item{seperator}{string}
}
\value{
data.frame
}
\description{
2 experimental conditions
}
