% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{processLFQ}
\alias{processLFQ}
\title{Process proteomics data, two experimental conditions}
\usage{
processLFQ(df, h0, h1, replicate = 3)
}
\arguments{
\item{df}{data.frame}

\item{h0}{string}

\item{h1}{string}

\item{replicate}{int}
}
\value{
data.frame
}
\description{
Process proteomics data, two experimental conditions
}
