% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI.R
\name{proteinDegree}
\alias{proteinDegree}
\title{calculate the degrees of proteins}
\usage{
proteinDegree(proteinID, species)
}
\arguments{
\item{proteinID}{char}

\item{species}{int}
}
\value{
dataframe
}
\description{
calculate the degrees of proteins
}
