% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUPred2A.R
\name{splitFasta}
\alias{splitFasta}
\title{split fasta files into smaller ones}
\usage{
splitFasta(inputFasta = file.choose(), x)
}
\arguments{
\item{inputFasta}{path}

\item{x}{int}
}
\description{
split fasta files into smaller ones
}
